% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatMisc.R
\name{CatMisc}
\alias{CatMisc}
\title{CAT Miscellaneous Functions}
\description{
A collection of utility methods frequently used in packages by
Charles Tilford (CAT)
}
\details{
\itemize{

  \item is.empty.field - Test for empty RefClass fields

  \item is.def - Tests a scalar against a variety of "nothing" values

  \item is.something - Tests if object is defined, and not zero, "" or FALSE

  \item parenRegExp - Simplifies capture from regular expression parentheses

  \item textBlockToVector - Splits a block of text into lines

  \item .flexFilehandle - Automatically handles .gz files

  \item methodHelp - Used inside RefClass methods for "self-help"
}
}
