% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefClassHelper.R
\name{colorNameToFunc}
\alias{colorNameToFunc}
\title{Color Name to Crayon Function}
\usage{
colorNameToFunc()
}
\value{
A list of lists
}
\description{
Function to generate a map of color names to crayon functions
}
\details{
The list generated by this method is simply a lookup list-of-lists
used to turn names ("cyan", "black") into \link[crayon]{crayon}
colorizing functions. It is utilized by \link{colorize} to convert
the user's color parameters into the appropriate colorizing
functions.

The function is run once, the first time the map is
requested. Afterwards the resulting list is stored in a local
variable for cached retrieval.
}
\seealso{
\link{colorize}
}
