% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatMisc.R
\name{.flexFilehandle}
\alias{.flexFilehandle}
\title{Flexible Filehandle}
\usage{
.flexFilehandle(file)
}
\arguments{
\item{file}{Required, the path to the file to read}
}
\value{
A list with filehandle ("fh"), basename ("name"), suffix
    ("sfx") and gzip flag ("gz")
}
\description{
Utility method to transparently handle 'normal' and compressed files
}
\details{
If the file ends with 'gz', then \code{gzfile} will be used to open
a file handle. Otherwise, \code{file} will be used.
}
