% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{dotSelfVarName}
\alias{dotSelfVarName}
\alias{.selfVarName}
\title{Get Name of Current Variable}
\arguments{
\item{def}{Default \code{"myObj"}, a default name that can be used
if the actual name can not be found.}

\item{fallbackVar}{Default \code{""}. Another default name, will be
used in preference to \code{def} if the real name can not be
found. There are ... reasons ... for this. I can't remember what
they were and am too timid to simplify the function at this point.}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
A single character string
}
\description{
For an object, get the variable name being used by R to 'hold' it
}
\details{
\preformatted{
## Method Usage:
myObject$.selfVarName( help=TRUE )

myObject$.selfVarName( def="myObj", fallbackVar="" )
}

Returns the variable name of an object. That is, if you have an
object \code{foo}, calling \code{foo$.selfVarName} should return
\code{"foo"}.

This functionality is used to generate self-referential help text
that can be copied-and-pasted to allow for easy exploration of the
object.
}
\examples{

rch <- RefClassHelper()
rch$.selfVarName()
identical("rch", rch$.selfVarName())
}
\seealso{
\link{varName}, \link{help}, \link{varCheck},
\link{getFieldDescriptions}
}
