% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{getFieldDescriptions}
\alias{getFieldDescriptions}
\title{Get Field Descriptions}
\arguments{
\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
A list of character strings
}
\description{
Get brief descriptions for all object fields
}
\details{
\preformatted{
## Method Usage:
myObject$getFieldDescriptions( help=TRUE )

myObject$getFieldDescriptions( )
}

Returns a list, with names representing object fields and values
being descriptions of each one. The descriptions need to be defined
in the object's \link{fieldDescriptions} method. If inheritance is
being used, they will be recovered recursively for each parent
class. The function is used by the \link{help} method.
}
\examples{

rch <- RefClassHelper()
rch$getFieldDescriptions()
rch$help() # Will be shown at the end of the help
}
\seealso{
\link{fieldDescriptions}, \link{getHelpSections},
\link{help}
}
