% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{getHelpSections}
\alias{getHelpSections}
\title{Get Help Sections}
\arguments{
\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
A list of character strings
}
\description{
Get a list of all notable object methods, organized into sections
}
\details{
\preformatted{
## Method Usage:
myObject$getHelpSections( help=TRUE )

myObject$getHelpSections( )
}

Returns a list, with names representing conceptual sections and
values being object methods assigned to those sections. This
structure needs to be defined in the object's \link{helpSections}
method. If inheritance is being used, sections will be recovered
recursively for each parent class. The function is used by the
\link{help} method.
}
\examples{

rch <- RefClassHelper()
rch$getHelpSections()
rch$help() # Will be used near the top to organize methods
}
\seealso{
\link{helpSections}, \link{getFieldDescriptions},
\link{help}
}
