% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{help}
\alias{help}
\title{Object Help}
\arguments{
\item{genericName}{Default 'myObject'. Used if R can not determine
the actual variable name of the object (via \link{.selfVarName})}

\item{color}{Default \code{NULL}, which will use the value defined
by \link{useColor}. If TRUE, then output will be colorized.}

\item{generic}{Default \code{FALSE}. If TRUE, then generic
Reference Class methods common to all RefClass objects will be
included.}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
NULL, invisibly
}
\description{
Summarize object methods and fields, and commands to get additional help
}
\examples{

rch <- RefClassHelper()
## Report help for the object
rch$help()
rch$help( generic=TRUE ) # Include generic methods
}
\seealso{
\link{getFieldDescriptions}, \link{getHelpSections},
\link{show}
}
