% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{helpSections}
\alias{helpSections}
\title{Help Sections}
\arguments{
\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
A list of character vectors, each containing one or more
methods within the section defined by that entry's name.
}
\description{
Get a list that organizes methods into conceptually-related sections
}
\details{
\preformatted{
## Method Usage:
myObject$helpSections( help=TRUE )

myObject$helpSections( )
}

This method simply returns a list that organizes the methods used
by the Reference Class object into related sections. It is intended
to provide self-documenting information to the \link{help}
function.

If the object inherits other RefClass classes, they will be
recursively queried for their own sections. This recursion is managed by
\link{getHelpSections}.
}
\examples{

el <- RefClassHelper()
el$helpSections()
el$help() # Methods will be broken out into the above sections
}
\seealso{
\link{getHelpSections}, \link{help}, \link{fieldDescriptions}
}
