% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatMisc.R
\name{is.def}
\alias{is.def}
\title{Is Defined}
\usage{
is.def(x)
}
\arguments{
\item{x}{The object to be tested}
}
\value{
TRUE if "defined", otherwise FALSE. The following objects
    are considered "not defined":

\itemize{

  \item \code{NULL}

  \item Objects that only contain \code{NA}

  \item Matrices with zero columns and zero rows

  \item Empty fields (TRUE value from \link{is.empty.field})

}
}
\description{
Broad test returning false for a variety of "empty" values
}
\details{
Born out of frustration with diverse possibilities for
"not present". The function is NOT vectorized, instead expecting
the passed object to be "a single thing".
}
\examples{

## TRUE (defined) objects
is.def(FALSE)
is.def(c(""))
noData <- matrix(numeric(), 0, 2,
           dimnames=list(sample=character(),attr=c("width","weight")))
is.def(noData) # No information, but it has two columns
is.def(c(NA, NA, "b"))

## FALSE (not defined)
is.def(NULL)
is.def(numeric())
is.def(c(NA,NA,NA,NA))
is.def(matrix(numeric(),0,0))  # 0x0 matrix

}
