% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatMisc.R
\name{methodHelp}
\alias{methodHelp}
\title{Method Help}
\usage{
methodHelp(mc, cl, inh)
}
\arguments{
\item{mc}{Required, the result of match.call, called just before
entering this function. This should allow automatic
determination of the method name, as well as the variable name
holding the object}

\item{cl}{Required, the class() of the object}

\item{inh}{The inherited packages, taken from .refClassDef@contains}
}
\value{
The mysterious 'help_files_with_topic' object R uses for
    managing internal help (which will be rendered if not
    captured), or NA if no topic could be found
}
\description{
Mechanism to identify relevant help topics from calling context
}
\details{
This function is part of an attempt to better document
ReferenceClass objects, here focusing on method documentation. The
function is designed to be called generically from within a method;
See Examples for a fleshed-out illustration. This allows richer
documentation to be maintained in Roxygen, and allows objects to be
self-documenting if the user passes a help=TRUE flag, eg:

\preformatted{
 myRefClass$methods(
   cube = function( x, help=FALSE ) {
       if (help) return( methodHelp(match.call(), class(.self),
                                    names(.refClassDef@contains)) )
       x ^ 3
   })
}
}
\examples{

mrct <- myRefClassThing(5)

# General information on the class as a whole:
?myRefClassThing

# Specific information on the $thingProduct() function
mrct$thingProduct( help=TRUE )

}
