% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefClassHelper.R
\docType{class}
\name{myRefClassStuff-class}
\alias{myRefClassStuff-class}
\alias{myRefClassStuff}
\title{Example Inherited RefClass Object}
\description{
A tiny object designed to be inherited by another tiny object
}
\details{
This is a toy ReferenceClass (aka 'R5') object that is used by
another toy class (\link{myRefClassThing}), and is used for tests
of \link{allRefClasses}.
}
\section{Fields}{

\describe{
\item{\code{txt}}{An interesting text value}
}}

\section{Methods}{

\describe{
\item{\code{initialize(useColor = TRUE, ...)}}{Create a new RefClassHelper Reference Class object}

\item{\code{thingText(prefix = "[Thing Text]", help = FALSE)}}{Reports the text held by the thing}
}}

\seealso{
\link{myRefClassThing}, \link{allRefClasses}
}
