% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatMisc.R
\docType{class}
\name{myRefClassThing-class}
\alias{myRefClassThing-class}
\alias{myRefClassThing}
\title{Example Reference Class Object}
\description{
A tiny object that multiplies things
}
\details{
This is a toy ReferenceClass (aka 'R5') object used to illustrate
the \link{methodHelp} function's use in documenting object
methods. It is also utilized by tests for \code{methodHelp}.
}
\section{Fields}{

\describe{
\item{\code{x}}{A numeric value}
}}

\section{Methods}{

\describe{
\item{\code{thingProduct(y = 7, help = FALSE)}}{Multiplies the x field by parameter y}
}}

\examples{

# Help at the class level
myRefClassThing( help=TRUE )

x <- myRefClassThing( x=17 )
# Help at the method level
x$thingProduct( help=TRUE )

}
\seealso{
\link{methodHelp}
}
