% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefClassHelper.R
\docType{class}
\name{RefClassHelper-class}
\alias{RefClassHelper-class}
\alias{RefClassHelper}
\title{Reference Class Helper Class}
\description{
A utility class that provides help functions when inherited
}
\details{
This class is not designed to be used on its own, but rather to be
inherited by another Reference Class.
}
\section{Fields}{

\describe{
\item{\code{useCol}}{Default TRUE. Logical flag to indicate if color
should be used in messaging}

\item{\code{varName}}{Extracted variable name of the object}

\item{\code{varCheck}}{Last time an attempt was made to extract varName}
}}

\section{Methods}{

\describe{
\item{\code{annotateFields(help = FALSE)}}{Update object fields to include attributes with brief descriptions}

\item{\code{colorize(msg = "", color = NULL, bgcolor = NULL, help = FALSE)}}{Use crayon to add ANSI color codes to text}

\item{\code{colorMap(color, bg = FALSE, help = FALSE)}}{Convert a color name into a crayon colorizing function}

\item{\code{fieldDescriptions(help = FALSE)}}{A static list of brief descriptions for each field in this object}

\item{\code{getFieldDescriptions(help = FALSE)}}{Recursively process an object's fieldDescriptions() return value}

\item{\code{help(color = NULL, help = FALSE)}}{Display high-level help about all object methods}

\item{\code{initialize(useColor = TRUE, ...)}}{Create a new RefClassHelper Reference Class object}

\item{\code{showHelp(sections = list(), genericName = "myObject", color = NULL,
  generic = FALSE, help = FALSE)}}{Construct text for the help() method}

\item{\code{useColor(newval = NULL, help = FALSE)}}{Get or set the flag determining if messages are colorized}
}}

