% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatMisc.R
\name{textBlockToVector}
\alias{textBlockToVector}
\title{Text Block to Vector}
\usage{
textBlockToVector(x, split = "[\\n\\r]", trim.white = TRUE,
  skip.empty = TRUE)
}
\arguments{
\item{x}{The block of text to parse}

\item{split}{Default \code{"[\n\r]+"}, the regular expression used
to split out lines}

\item{trim.white}{Default TRUE, which will cause leading and
trailing whitespace to be removed from each line}

\item{skip.empty}{Default TRUE, which will remove elements that are
the empty string (\code{''})}
}
\value{
A character vector, each element being a line
}
\description{
Convert a multi-line string to a vector of lines
}
\details{
When embedding static lists of text in code I find it sometimes
easier to read and maintain the list if it is encoded as a simple
block of text. This function will break such a block into a vector
}
\examples{

myBlock <- "
A quick brown fox
  Five golden rings  
Klaatu barada nikto
"

textBlockToVector(myBlock)

}
