% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatMisc.R
\name{thingProduct}
\alias{thingProduct}
\title{Thing Product}
\arguments{
\item{y}{The second number}
}
\value{
A numeric product of x * y
}
\description{
A toy object method used to illustrate \link{methodHelp}
}
\details{
Multiplies the internally-stored value of x by a supplied
    second number.
}
\examples{

mrct <- myRefClassThing(5)
mrct$thingProduct(11)

}
