% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefClassHelper.R
\name{thingText}
\alias{thingText}
\title{Thing Text}
\arguments{
\item{prefix}{Default '[Thing Text]'. A prefix to include before
the actual Thing text.}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
A character vector
}
\description{
Get text for toy (demonstration) Reference Class object
}
\details{
\preformatted{## Method Usage:
myObject$thingText( help=TRUE )

myObject$thingText( prefix="[Thing Text]" )
}

Returns the thing's text, with a prefix
}
\examples{

mrcs <- myRefClassStuff(txt="Some text for the thing")
## Query the text field directly
mrcs$txt
## Access the text by this method
mrcs$thingText(prefix="[Hooray!]")
# Method from the parent class (RefClassHelper)
message(mrcs$colorize("This text is blue", color='blue'))

}
