% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{useCol}
\alias{useCol}
\title{RefClassHelper Use Color Flag}
\value{
A single logical value
}
\description{
Internal RefClass field indicating if output should be colorized
}
\details{
A logical flag, if TRUE it indicates that messages (from compliant
methods) should be colorized with the \link[crayon]{crayon}
package.

\preformatted{
## NORMALLY YOU WILL NOT WANT TO ACCESS THIS FIELD DIRECTLY
## Instead, use the \link{useColor} method to check and alter the value
}
}
\examples{

rch <- RefClassHelper( )
message(rch$colorize("I'm feeling blue", "blue"))
rch$useColor(FALSE)
message(rch$colorize("I'm still down, I'm just not showing it", "blue"))
}
\seealso{
\link{useColor}, \link{message}
}
