% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{varName}
\alias{varName}
\title{RefClassHelper Variable Name}
\description{
Internal RefClass field holding its own variable name
}
\details{
A character string containing the name of the object. That is, an
object invoked with the R variable \code{x} should have "x" as the
value for this field. It is set using the \link{.selfVarName}
function.
}
\examples{

rch <- RefClassHelper( )
rch$varName # Should initially be unset
rch # Pretty-print the object, which will attempt to discover its name
rch$varName # Should now be "rch"
}
\seealso{
\link{.selfVarName}, \link{help}, \link{varCheck},
\link{getFieldDescriptions}
}
