% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefClassHelper.R
\docType{class}
\name{myRefClassThing-class}
\alias{myRefClassThing-class}
\alias{myRefClassThing}
\title{Example Reference Class Object}
\description{
A tiny object that multiplies things
}
\details{
This is a toy ReferenceClass (aka 'R5') object used to illustrate
the \link{methodHelp} function's use in documenting object
methods. It is also utilized by tests for \code{methodHelp}.
}
\section{Fields}{

\describe{
\item{\code{x}}{A numeric value}
}}

\section{Methods}{

\describe{
\item{\code{fieldDescriptions(help = FALSE)}}{A static list of brief descriptions for each field in this object}

\item{\code{helpSections(help = FALSE)}}{Static list organizing object methods into conceptual sections}

\item{\code{initialize(txt = "Initial text", ...)}}{Initialize a new myRefClassStuff object}

\item{\code{show(...)}}{Provide pretty-printed summary for RefClassStuff objects}

\item{\code{thingProduct(y = 7, help = FALSE)}}{Multiplies the x field by parameter y}
}}

\examples{

mrct <- myRefClassThing( x=17, txt="Text is managed by Stuff" )
# Help at the object level:
mrct$help()
# Help at the method level:
mrct$thingProduct( help=TRUE )

# This class's method:
mrct$thingProduct(3)
# Method from the parent class (myRefClassStuff):
mrct$thingText(prefix="[Foo]")
# Method from the grandparent class (RefClassHelper)
message(mrct$colorize("This text is red", color='red'))

}
\seealso{
\link{methodHelp}
}
