% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefClassHelper.R
\name{thingProduct}
\alias{thingProduct}
\title{Thing Product}
\arguments{
\item{y}{The second number}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
A numeric product of x * y
}
\description{
A toy object method used to illustrate \link{methodHelp}
}
\details{
\preformatted{ ## Method Usage:
myObject$thingProduct( help=TRUE )

myObject$thingProduct( y=7 )
}

Multiplies the internally-stored value of x by a supplied second
number.
}
\examples{

mrct <- myRefClassThing(5)
mrct$thingProduct(11)

}
